requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.7.0
boto3==1.34.54
botocore==1.34.54
bs4==0.0.2
CacheControl==0.14.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
dataclasses-json==0.6.4
Django==5.0.4
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
frozenlist==1.4.1
grpcio==1.62.1
grpcio-health-checking==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
ibm-cos-sdk==2.13.4
ibm-cos-sdk-core==2.13.4
ibm-cos-sdk-s3transfer==2.13.4
ibm-watson-machine-learning==1.0.354
ibm-watsonx-ai==0.2.4
idna==3.6
importlib_metadata==7.1.0
incremental==22.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.6
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-ibm==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.41
lomond==0.3.3
marshmallow==3.21.1
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.0
packaging==23.2
pandas==2.1.4
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.1
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.3
Twisted==24.3.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.0.7
validators==0.22.0
vine==5.1.0
wcwidth==0.2.13
weaviate-client==4.5.5
yarl==1.9.4
zipp==3.18.1
zope.interface==6.2
